// apps/web/src/lib/api/generated/types/groupMemberResponseDtoRole.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * Movies App API
 * REST API for Movies application

**Links:**
- [OpenAPI JSON](/api/docs/json)
- [OpenAPI YAML](/api/docs/yaml)
 * OpenAPI spec version: 1.0.0
 */

export type GroupMemberResponseDtoRole = 'admin' | 'moderator' | 'member';